Update src/main/java/htmlpublisher/HtmlPublisherTarget.java
1.The Job level link of HTML report is modified to be the last build’s result, 
not the last successful build’s result(line 155). According to my team's feedback,
most of them thought the link in Job page is the last report, and they stopped using
this link when they could not find out what report it links to. 
2.Discard the redundant link on the Job level page. I find that the link in side panel
 and the link in Job summary link to the same report. I think it is not necessary to 
put two same link in same page.  
  