adding backToName and backToUrl variables
When viewing HTML Report connected to build it is convenient to be
able to use "Back To ..." option to return to build not to the project.
This commit adds additional methods to BaseHTMLAction and overrides in
HTMLBuildAction. Thanks to that we can easily go back to project or to
build accordingly.

Signed-off-by: Artur Harasimiuk <artur.harasimiuk@intel.com>