Allow overriding Content Security Policy on published artefacts
Jenkins currently enforces a global Content Security Policy that
prevents in-line styles and scripts from being loaded, so prevents
archived HTML from tools that use such in-line elements from rendering
properly. The workaround for this is to weaken the global policy by
allowing such in-line elements to be loaded across Jenkins, although in
some places people are recommending to disable CSP completely. To
overcome this, the plugin is being altered to allow configuration to be
set that enables scripts and styles, and dynamically injects these into
the Content-Security-Policy header when serving archived artefacts.

To ensure that a user with Job Configure permissions can't initially use
their access to enable a Content Security Policy to publish malicious
artefacts for attacking the Jenkins instance, the option to alter the
Content Security Policy is enabled instance-wide through a global
configuration by default, and must be enabled through the Jenkins UI by
an Administrator.